+++ SemiPlot/SemiPlot.Core/Trends/Sample.cs
@@ -1,3 +1,3 @@
 ﻿namespace SemiPlot.Core.Trends;
 
-public sealed record Sample(long PenId, DateTime TimestampUtc, double Value);
+public sealed record Sample(int PenId, DateTime TimestampUtc, double Value);
