+++ SemiPlot/SemiPlot.Tests/Core/Trends/PenScaleModelTests.cs
@@ -22,7 +22,7 @@ public sealed class PenScaleModelTests
 			new PenScaleSettings(PenId: 1, AxisKey: "pressure"),
 			new PenScaleSettings(PenId: 2, AxisKey: "temperature")
 		};
-		var envelopes = new Dictionary<long, PenHistoryEnvelope>
+		var envelopes = new Dictionary<int, PenHistoryEnvelope>
 		{
 			[1] = Envelope(1, (0.0, 10.0), (2.0, 12.0)),
 			[2] = Envelope(2, (100.0, 200.0))
@@ -48,7 +48,7 @@ public sealed class PenScaleModelTests
 			new PenScaleSettings(PenId: 1, AxisKey: "a"),
 			new PenScaleSettings(PenId: 2, AxisKey: "b")
 		};
-		var envelopes = new Dictionary<long, PenHistoryEnvelope>
+		var envelopes = new Dictionary<int, PenHistoryEnvelope>
 		{
 			[1] = Envelope(1, (0.0, 5.0)),
 			[2] = Envelope(2, (-3.0, 7.0))
@@ -59,7 +59,7 @@ public sealed class PenScaleModelTests
 		scales.Should().HaveCount(2);
 		var axisB = scales.Single(scale => scale.AxisKey == "b");
 		axisB.IsActive.Should().BeFalse();
-		axisB.PenIds.Should().ContainSingle().Which.Should().Be(2L);
+		axisB.PenIds.Should().ContainSingle().Which.Should().Be(2);
 		axisB.Min.Should().BeApproximately(-3.0 - (10.0 * 0.05), 1e-9);
 		axisB.Max.Should().BeApproximately(7.0 + (10.0 * 0.05), 1e-9);
 	}
@@ -74,7 +74,7 @@ public sealed class PenScaleModelTests
 			new PenScaleSettings(PenId: 2, AxisKey: "heaters"),
 			new PenScaleSettings(PenId: 3, AxisKey: "heaters")
 		};
-		var envelopes = new Dictionary<long, PenHistoryEnvelope>
+		var envelopes = new Dictionary<int, PenHistoryEnvelope>
 		{
 			[1] = Envelope(1, (10.0, 20.0)),
 			[2] = Envelope(2, (5.0, 30.0)),
@@ -85,7 +85,7 @@ public sealed class PenScaleModelTests
 
 		var shared = scales.Should().ContainSingle().Which;
 		shared.AxisKey.Should().Be("heaters");
-		shared.PenIds.Should().BeEquivalentTo(new[] { 1L, 2L, 3L });
+		shared.PenIds.Should().BeEquivalentTo(new[] { 1, 2, 3 });
 		shared.Min.Should().BeApproximately(5.0 - (25.0 * 0.05), 1e-9);
 		shared.Max.Should().BeApproximately(30.0 + (25.0 * 0.05), 1e-9);
 	}
@@ -100,7 +100,7 @@ public sealed class PenScaleModelTests
 		var min = new[] { 0.0, 50.0, 1000.0, -500.0 };
 		var max = new[] { 1.0, 60.0, 2000.0, -400.0 };
 		var center = new[] { 0.5, 55.0, 1500.0, -450.0 };
-		var envelopes = new Dictionary<long, PenHistoryEnvelope>
+		var envelopes = new Dictionary<int, PenHistoryEnvelope>
 		{
 			[1] = new PenHistoryEnvelope(1, timestamps, min, max, center)
 		};
@@ -121,7 +121,7 @@ public sealed class PenScaleModelTests
 		{
 			new PenScaleSettings(PenId: 1, AxisKey: "a", Mode: ScaleMode.Manual, ManualMin: -2.0, ManualMax: 8.0)
 		};
-		var envelopes = new Dictionary<long, PenHistoryEnvelope> { [1] = Envelope(1, (1000.0, 2000.0)) };
+		var envelopes = new Dictionary<int, PenHistoryEnvelope> { [1] = Envelope(1, (1000.0, 2000.0)) };
 
 		var scales = model.Compute(settings, envelopes, activePenId: 1, _origin, _origin.AddHours(1));
 
@@ -141,7 +141,7 @@ public sealed class PenScaleModelTests
 		var min = new[] { -5.0, 2.0, 0.0 };
 		var max = new[] { 0.0, 4.0, 50.0 };
 		var center = new[] { -2.0, 3.0, 25.0 };
-		var envelopes = new Dictionary<long, PenHistoryEnvelope>
+		var envelopes = new Dictionary<int, PenHistoryEnvelope>
 		{
 			[1] = new PenHistoryEnvelope(1, timestamps, min, max, center)
 		};
@@ -159,7 +159,7 @@ public sealed class PenScaleModelTests
 	{
 		var model = new PenScaleModel();
 		var settings = new[] { new PenScaleSettings(PenId: 1, AxisKey: "a", IsLogarithmic: true) };
-		var envelopes = new Dictionary<long, PenHistoryEnvelope> { [1] = Envelope(1, (-10.0, -1.0)) };
+		var envelopes = new Dictionary<int, PenHistoryEnvelope> { [1] = Envelope(1, (-10.0, -1.0)) };
 
 		var scales = model.Compute(settings, envelopes, activePenId: 1, _origin, _origin.AddHours(1));
 
@@ -173,7 +173,7 @@ public sealed class PenScaleModelTests
 	{
 		var model = new PenScaleModel();
 		var settings = new[] { new PenScaleSettings(PenId: 1, AxisKey: "a", IsVisible: false) };
-		var envelopes = new Dictionary<long, PenHistoryEnvelope> { [1] = Envelope(1, (0.0, 10.0)) };
+		var envelopes = new Dictionary<int, PenHistoryEnvelope> { [1] = Envelope(1, (0.0, 10.0)) };
 
 		var scales = model.Compute(settings, envelopes, activePenId: 1, _origin, _origin.AddHours(1));
 
@@ -190,7 +190,7 @@ public sealed class PenScaleModelTests
 		var min = new[] { 4.0, double.NaN, 6.0 };
 		var max = new[] { 8.0, double.NaN, 10.0 };
 		var center = new[] { 6.0, double.NaN, 8.0 };
-		var envelopes = new Dictionary<long, PenHistoryEnvelope>
+		var envelopes = new Dictionary<int, PenHistoryEnvelope>
 		{
 			[1] = new PenHistoryEnvelope(1, timestamps, min, max, center)
 		};
@@ -208,7 +208,7 @@ public sealed class PenScaleModelTests
 	{
 		var model = new PenScaleModel();
 		var settings = new[] { new PenScaleSettings(PenId: 1, AxisKey: "a") };
-		var envelopes = new Dictionary<long, PenHistoryEnvelope> { [1] = Envelope(1, (5.0, 5.0)) };
+		var envelopes = new Dictionary<int, PenHistoryEnvelope> { [1] = Envelope(1, (5.0, 5.0)) };
 
 		var scales = model.Compute(settings, envelopes, activePenId: 1, _origin, _origin.AddHours(1));
 
@@ -225,7 +225,7 @@ public sealed class PenScaleModelTests
 		{
 			new PenScaleSettings(PenId: 1, AxisKey: "a", Mode: ScaleMode.Manual, ManualMin: 90.0, ManualMax: 10.0)
 		};
-		var envelopes = new Dictionary<long, PenHistoryEnvelope> { [1] = Envelope(1, (0.0, 1.0)) };
+		var envelopes = new Dictionary<int, PenHistoryEnvelope> { [1] = Envelope(1, (0.0, 1.0)) };
 
 		var scales = model.Compute(settings, envelopes, activePenId: 1, _origin, _origin.AddHours(1));
 
@@ -243,7 +243,7 @@ public sealed class PenScaleModelTests
 			new PenScaleSettings(
 				PenId: 1, AxisKey: "a", Mode: ScaleMode.Manual, ManualMin: -5.0, ManualMax: 100.0, IsLogarithmic: true)
 		};
-		var envelopes = new Dictionary<long, PenHistoryEnvelope> { [1] = Envelope(1, (1.0, 50.0)) };
+		var envelopes = new Dictionary<int, PenHistoryEnvelope> { [1] = Envelope(1, (1.0, 50.0)) };
 
 		var scales = model.Compute(settings, envelopes, activePenId: 1, _origin, _origin.AddHours(1));
 
@@ -252,7 +252,7 @@ public sealed class PenScaleModelTests
 		manual.Max.Should().Be(100.0);
 	}
 
-	private static PenHistoryEnvelope Envelope(long penId, params (double Min, double Max)[] columns)
+	private static PenHistoryEnvelope Envelope(int penId, params (double Min, double Max)[] columns)
 	{
 		var timestamps = new DateTime[columns.Length];
 		var min = new double[columns.Length];
